+++ Helper/Helper/CommonViews/ContentView.swift
@@ -20,7 +20,7 @@ struct ContentView: View {
     // shared for next view
     static let shared = ContentView()
     
-    // Constants for location
+    // Constants for location for getting current location
     var locationManager = LocationManager()
     
     // fetching user data from core data
@@ -41,14 +41,6 @@ struct ContentView: View {
         } catch {
             print(error)
         }
-//        do {
-//            results.forEach{ (user) in
-//                context.delete(user)
-//            }
-//            try context.save()
-//        } catch {
-//            print(error)
-//        }
     }
     
     
@@ -63,24 +55,19 @@ struct ContentView: View {
 //                ProgressView().onAppear(perform: {clearData(entityName: "Task")})
 ////                ProgressView().onAppear(perform: {taskModel.fetchData(context: context)})
 //            }
-            let _ = print("language is: \(Locale.preferredLanguages[0])")
-            // checking if core data exists
+            
+            // checking if core data exists, if not fetch data from Network
             if results.isEmpty {
                 if userModel.users.isEmpty {
                     ProgressView().onAppear(perform: {userModel.fetchData(context: context)})
                 }
-            } else {
-                let _ = print("read user from core \(results.count) \(String(describing: results[results.count-1].email)) \(String(describing: results[results.count-1].password))")
-                let _ = print("user \(results)")
             }
+            
             if taskResults.isEmpty {
                 if taskModel.tasks.isEmpty {
                     ProgressView().onAppear(perform: {taskModel.fetchData(context: context)})
                 }
-            } else {
-                let _ = print("read task from core \(taskResults.count) \(String(describing: taskResults[taskResults.count-1].title)) \(String(describing: taskResults[taskResults.count-1].time))")
             }
-            
         }
         LandingPage()
     }
@@ -89,36 +76,5 @@ struct ContentView: View {
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
-    }
-}
-
-final class LocationViewModel: NSObject, ObservableObject, CLLocationManagerDelegate{
-
-    @Published var region = MKCoordinateRegion(center: CLLocationCoordinate2D(latitude: 40, longitude: 120), span: MKCoordinateSpan(latitudeDelta: 100, longitudeDelta: 100))
-
-    let locationManager = CLLocationManager()
-
-    override init() {
-        super.init()
-        locationManager.requestAlwaysAuthorization()
-        locationManager.requestWhenInUseAuthorization()
-        locationManager.delegate = self
-        locationManager.desiredAccuracy = kCLLocationAccuracyNearestTenMeters
-        locationManager.startUpdatingLocation()
-    }
-    func requestAllowOnceLocationPermission(){
-        locationManager.requestLocation()
-    }
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let latestLocation = locations.last else {
-            return
-        }
-        DispatchQueue.main.async {
-            self.region = MKCoordinateRegion(center: latestLocation.coordinate, span: MKCoordinateSpan(latitudeDelta: 0.01, longitudeDelta: 0.01))
-            print("location is \(latestLocation.coordinate)")
-        }
-    }
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        print(error.localizedDescription)
     }
 }
