+++ Helper/Helper/Manager/LocationManager.swift
@@ -24,7 +24,7 @@ class LocationManager: NSObject {
     public var exposedLocation: CLLocation? {
         return self.locationManager.location
     }
-        
+    
     override init() {
         super.init()
         self.locationManager.delegate = self
@@ -39,7 +39,7 @@ extension LocationManager: CLLocationManagerDelegate {
     
     // check permission for using location
     func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-
+        
         switch status {
             case .notDetermined         : print("location permission notDetermined")        // location permission not asked for yet
             case .authorizedWhenInUse   : print("location permission authorizedWhenInUse")  // location authorized
@@ -57,7 +57,6 @@ extension LocationManager: CLLocationManagerDelegate {
     
     // get current location
     func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        print("location update:: \(locations[0])")
         let geocoder = CLGeocoder()
         geocoder.reverseGeocodeLocation(locations[0]) { placemarks, error in
             guard error == nil else {
@@ -69,38 +68,37 @@ extension LocationManager: CLLocationManagerDelegate {
                 print("*** Error in \(#function): placemark is nil")
                 return
             }
-
-            if placemark != nil {
-                self.address = placemark.name! + ", " + placemark.postalCode! + " " + placemark.locality! + ", " + placemark.country!
-                self.lat = (placemark.location?.coordinate.latitude)!
-                self.long = (placemark.location?.coordinate.longitude)!
-            }
+            
+            self.address = placemark.name! + ", " + placemark.postalCode! + " " + placemark.locality! + ", " + placemark.country!
+            self.lat = (placemark.location?.coordinate.latitude)!
+            self.long = (placemark.location?.coordinate.longitude)!
+            
         }
     }
     
     // get location from address
     func getLocation(forPlaceCalled name: String, completion: @escaping(CLLocation?) -> Void) {
+        
+        let geocoder = CLGeocoder()
+        geocoder.geocodeAddressString(name) { placemarks, error in
             
-            let geocoder = CLGeocoder()
-            geocoder.geocodeAddressString(name) { placemarks, error in
-                
-                guard error == nil else {
-                    print("* Error in \(#function): \(error!.localizedDescription)")
-                    completion(nil)
-                    return
-                }
-                
-                guard let placemark = placemarks?[0] else {
-                    print("* Error in \(#function): placemark is nil")
-                    completion(nil)
-                    return
-                }
-                
-                guard let location = placemark.location else {
-                    print("* Error in \(#function): placemark is nil")
-                    return
-                }
-                completion(location)
+            guard error == nil else {
+                print("* Error in \(#function): \(error!.localizedDescription)")
+                completion(nil)
+                return
             }
+            
+            guard let placemark = placemarks?[0] else {
+                print("* Error in \(#function): placemark is nil")
+                completion(nil)
+                return
+            }
+            
+            guard let location = placemark.location else {
+                print("* Error in \(#function): placemark is nil")
+                return
+            }
+            completion(location)
         }
+    }
 }
