+++ Helper/Helper/Manager/fi.lproj/Localizable.strings
@@ -104,3 +104,9 @@
 "Thanks for your help!" = "Kiitos avusta!";
 "Check completed task in history menu" = "Tarkista suoritettu tehtävä historiavalikosta";
 "Your location" = "Sijaintisi";
+"Already have an account?" = "Onko sinulla jo tili?";
+"CALL" = "SOITTAA";
+"You don't have any task" = "Sinulla ei ole mitään tehtävää";
+"You didnot do any task" = "Et tehnyt mitään tehtävää";
+"No task is avaiable" = "Tehtävää ei ole käytettävissä";
+"New Task" = "Uusi tehtävä";
