+++ Helper/Helper/VolunteerViews/Location.swift
@@ -11,24 +11,6 @@ import MapKit
 
 struct Location: View {
     
-    //    @IBOutlet weak var mapView: MKMapView!
-    //
-    //    @FetchRequest(entity: User.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \User.userId, ascending: true)]) var results: FetchedResults<User>
-    //
-    //    override func viewDidLoad() {
-    //        super.viewDidLoad()
-    //        createAnnotations(users: results)
-    //    }
-    //
-    //    func createAnnotations(users: FetchedResults<User>) {
-    //        for user in users {
-    //            let annotations = MKPointAnnotation()
-    //            annotations.title = user.fullname
-    //            annotations.coordinate = CLLocationCoordinate2D(latitude: user.lat, longitude: user.long)
-    //            mapView.addAnnotation(annotations)
-    //        }
-    //    }
-    
     @Binding var volunteerName: String
     
     var body: some View {
@@ -49,26 +31,23 @@ struct Location: View {
                         .multilineTextAlignment(.center)
                         .foregroundColor(Color("Primary"))
                         .padding(.all, 20)
-                    NavigationLink(
-                        destination: MapView(volunteerName: $volunteerName).navigationBarBackButtonHidden(true),
-                        label: {
-                            HStack {
-                                Image(systemName: "location.fill")
-                                    .foregroundColor(.white)
-                                Text("Go To Map")
-                                    .bold()
-                                    .foregroundColor(.white)
-                            }
-                            .frame(width: 280, height: 60, alignment: .center)
-                            .background(Color("Primary"))
-                            .cornerRadius(5)
-                        }).padding(.top, 10)
+                    NavigationLink(destination: MapView(volunteerName: $volunteerName).navigationBarBackButtonHidden(true), label: {
+                        HStack {
+                            Image(systemName: "location.fill")
+                                .foregroundColor(.white)
+                            Text("Go To Map")
+                                .bold()
+                                .foregroundColor(.white)
+                        }
+                        .frame(width: 280, height: 60, alignment: .center)
+                        .background(Color("Primary"))
+                        .cornerRadius(5)
+                    })
+                    .padding(.top, 10)
                 }
             }
         }
-        
-    }
-    
+    }    
 }
 
 struct Location_Previews: PreviewProvider {
